+++ internal/report/pretty.go
@@ -31,6 +31,6 @@ func (r *Pretty) ActionDone(id string, _ driver.Output, duration time.Duration, 
 	console.Printlnf("  ✓ %s (%dms)", id, duration.Milliseconds())
 }
 
-func (r *Pretty) ActionFail(id string, err error, duration time.Duration, _ map[string]any) {
+func (r *Pretty) ActionFail(id string, _ driver.Output, err error, duration time.Duration, _ map[string]any) {
 	console.Printlnf("  ✗ %s (%dms): %s", id, duration.Milliseconds(), err)
 }