+++ internal/report/reporter.go
@@ -11,5 +11,5 @@ type Reporter interface {
 	FlowDone(name string, state map[string]any)
 	ActionStart(id, description string)
 	ActionDone(id string, output driver.Output, duration time.Duration, state map[string]any)
-	ActionFail(id string, err error, duration time.Duration, state map[string]any)
+	ActionFail(id string, output driver.Output, err error, duration time.Duration, state map[string]any)
 }